Make withStructuredOutput validate if passed a Zod schema
